docs: 커스텀 메서드, 타입 주석추가
