+++ Projects/Core/Sources/ViewController/PageViewController/Extension+Rx/BasePageVC+Rx.swift
@@ -10,6 +10,9 @@ import RxSwift
 import RxCocoa
 
 public extension Reactive where Base: BasePageViewController {
+    
+    /// Reactive Rapper for `CurrentPage`Property
+    /// You can assign a value to change the `CurrentPage`property's value through this value
     var goToPage: Binder<Int> {
         return Binder(self.base) { pageVC, index in
             base.goToPage(index)
