+++ Projects/Modules/DSKit/Sources/Components/View/ImageView/ImageView_Honey.swift
@@ -8,6 +8,10 @@
 
 import UIKit
 
+/// 프로젝트에 사용되는 꿀통 ImageView입니다.
+/// 생성 후 setHoney 메서드를 호출해야 이미지가 세팅됩니다.
+/// 생성자에 지정한 크기대로 컴포넌트의 크기가 설정됩니다.
+/// 크기에 대한 오토레이아웃을 따로 지정하지 않아도 됩니다.
 public final class ImageView_Honey: UIImageView {
     
     public enum HoneyType {
@@ -96,6 +100,9 @@ public final class ImageView_Honey: UIImageView {
     private let type: HoneyType
     private var temparature: Double?
     
+    /// 꿀통 이미지뷰를 생성하는 생성자입니다.
+    /// type은 large, small 두 가지가 있습니다.
+    /// type은 피그마에 해당 컴포넌트를 클릭하면 나타납니다.
     public init(type: HoneyType) {
         self.type = type
         super.init(frame: .zero)
