--- Projects/Modules/DSKit/Sources/Components/View/Label/PaddingLabel.swift
@@ -1,9 +1,0 @@
-//
-//  PaddingLabel.swift
-//  DSKit
-//
-//  Created by 강현준 on 2023/11/28.
-//  Copyright © 2023 TeamOne. All rights reserved.
-//
-
-import UIKit
