+++ Projects/Modules/DSKit/Sources/Extension/HandleError/UIViewController+ErrorAlert.swift
@@ -15,6 +15,8 @@ import SnapKit
 import Then
 
 public extension Reactive where Base: UIViewController {
+    /// Error를 띄웁니다.
+    /// Using: ViewController().rx.presentErrorAlert
     var presentErrorAlert: Binder<Error> {
         return Binder(self.base) { base, error in
             base.presentErrorAlert(
@@ -26,6 +28,13 @@ public extension Reactive where Base: UIViewController {
 
 public extension UIViewController {
     
+    /**
+     에러 알럿을 띄웁니다.
+     - Parameters:
+        - error: Error Type
+        - finishSubject: 에러가 사라지고 어떤 작업을 해야할 경우 사용합니다. 사용자가 확인을 누르면 True 이벤트가 전달됩니다.
+        - darkBackgound: 에러 알럿의 배경을 어둡게 처리할 건지 전달합니다
+     */
     func presentErrorAlert(
         error: Error?,
         finishSubject: PublishSubject<Bool>? = nil,
