+++ Projects/TeamOne/Sources/Setting/View/SettingMainView.swift
@@ -1,0 +1,9 @@
+//
+//  SettingMainView.swift
+//  TeamOne
+//
+//  Created by 강현준 on 1/30/24.
+//  Copyright © 2024 TeamOne. All rights reserved.
+//
+
+import Foundation
